pyproject.toml
[project]
name = "temporal-sheets-poc"
version = "0.1.0"
dependencies = [
    "temporalio>=1.0.0",
    "python-dotenv>=1.0.0",
    "google-api-python-client>=2.0.0",
    "google-auth-oauthlib>=1.0.0",
    "requests>=2.32.0",
    "jsonschema>=4.17.0",
    "colorlog>=6.0.0"
]
requires-python = ">=3.8"